Feature/EUI 3472 challenged access aa (#907)
* EUI-3472 squashed commit

EUI-3472 hardcoded placeholder

EUI-3472 pseudo guard and refinement of components

EUI-3472 feature toggling and trying to fix tests

* EUI-3472 trying to fix release merge

* EUI-3472 made it work with new logic

* EUI-3472 modified code to cater for specific access as well

* EUI-3472 fixed some issues

* EUI-3472 fix for functional tests

* EUI-3472 fix for duplication

* EUI-3472 linting

* EUI-3472 tests and refactoring

* EUI-3472 some refactoring

* EUI-3472 missing fields - unfinished

* EUI-3472 refactored mock and feature toggling, still need to fix locatio

* EUI-3472 some refactoring around court name

* EUI-3472 renamed variables

* EUI-3472 testing and refactoring

* EUI-3472 sonar

* EUI-3472 pre-merge acrobatics

* EUI-3472 post merge acrobatics

* EUI-3472 FormsModule fix

* EUI-3472 redundant code

* EUI-3472 managed to fix break

* EUI-3472 linting

* EUI-3472 trying to fix functional test

* EUI-3472 got court or hearing centre done

* EUI-3472 tag release version

* EUI-3472 fixed test and tag release

* EUI-3472 version linting

* EUI-3472 fixed oversight and new version

* EUI-3472 text fix

* EUI-3472 interface for AccessManagementBasicViewMock

* EUI-3472 removed double bangs

* EUI-3472 getting rid of useless sonar changes

* EUI-3472 new version

* EUI-3472 newer version as previous one failed functional tests

* EUI-3472 removed double bangs
