+++ demo/src/app/app.config.ts
@@ -30,7 +30,8 @@ export class AppConfig extends AbstractAppConfig {
     'work_allocation_api_url': '/workallocation',
     'user_info_api_url': '/user-info',
     'document_management_url_v2': '/v2/health',
-    'document_management_secure_enabled':true
+    'document_management_secure_enabled':true,
+    'access_management_mode':true
 
   };
 
@@ -155,5 +156,9 @@ export class AppConfig extends AbstractAppConfig {
 
   public getRemotePrintServiceUrl(): string {
     return this.config.print_service_url;
+  }
+
+  public getAccessManagementMode(): boolean {
+    return this.config.access_management_mode;
   }
 }
