+++ src/shared/components/case-editor/case-edit-page/case-edit-page.component.ts
@@ -18,6 +18,7 @@ import { Wizard } from '../domain/wizard.model';
 import { CaseField } from '../../../domain/definition';
 import { FieldsUtils } from '../../../services/fields';
 import { CaseFieldService } from '../../../services/case-fields/case-field.service';
+import { initDialog } from '../../helpers';
 
 @Component({
   selector: 'ccd-case-edit-page',
@@ -76,7 +77,7 @@ export class CaseEditPageComponent implements OnInit, AfterViewChecked {
   ) {}
 
   public ngOnInit(): void {
-    this.initDialog();
+    initDialog(this.dialogConfig);
     this.eventTrigger = this.caseEdit.eventTrigger;
     this.editForm = this.caseEdit.form;
     this.wizard = this.caseEdit.wizard;
@@ -308,21 +309,6 @@ export class CaseEditPageComponent implements OnInit, AfterViewChecked {
     return this.eventTrigger && this.eventTrigger.can_save_draft
       ? CaseEditPageComponent.TRIGGER_TEXT_SAVE
       : CaseEditPageComponent.TRIGGER_TEXT_START
-  }
-
-  private initDialog() {
-    this.dialogConfig = new MatDialogConfig();
-    this.dialogConfig.disableClose = true;
-    this.dialogConfig.autoFocus = true;
-    this.dialogConfig.ariaLabel = 'Label';
-    this.dialogConfig.height = '245px';
-    this.dialogConfig.width = '550px';
-    this.dialogConfig.panelClass = 'dialog';
-
-    this.dialogConfig.closeOnNavigation = false;
-    this.dialogConfig.position = {
-      top: window.innerHeight / 2 - 120 + 'px', left: window.innerWidth / 2 - 275 + 'px'
-    }
   }
 
   private discard(): void {
