+++ src/shared/components/case-viewer/case-basic-access-view/case-basic-access-view.component.html
@@ -1,0 +1,36 @@
+<form>
+  <div>
+    <cut-alert type="information">This case requires {{accessType | lowercase}} access.<br />
+      <ng-container *ngIf="accessType === 'CHALLENGED'">This is because the case is outside your work area. If you request access to the case, it will be logged for auditing purposes.</ng-container>
+      <ng-container *ngIf="accessType === 'SPECIFIC'">This means it's outside your jurisdiction and requires authorisation. If you request access to the case, it will be logged for auditing purposes.</ng-container>
+    </cut-alert>
+    <h1 class="heading-h1">#{{ caseDetails.case_id | ccdCaseReference}}</h1>
+    <h2 class="govuk-heading-m">Case details</h2>
+    <dl class="govuk-summary-list">
+      <div class="govuk-summary-list__row" *ngIf="caseDetails.basicFields?.caseNameHmctsInternal">
+        <dt class="govuk-summary-list__key summary-key">Case name</dt>
+        <dd class="govuk-summary-list__value">{{caseDetails.basicFields.caseNameHmctsInternal}}</dd>
+      </div>
+      <div class="govuk-summary-list__row" *ngIf="caseDetails.case_type?.jurisdiction?.name">
+        <dt class="govuk-summary-list__key summary-key">Service</dt>
+        <dd class="govuk-summary-list__value">{{caseDetails.case_type.jurisdiction.name}}</dd>
+      </div>
+      <div class="govuk-summary-list__row" *ngIf="caseDetails.state?.description">
+        <dt class="govuk-summary-list__key summary-key">State</dt>
+        <dd class="govuk-summary-list__value">{{caseDetails.state.description}}</dd>
+      </div>
+      <div class="govuk-summary-list__row" *ngIf="courtOrHearingCentre">
+        <dt class="govuk-summary-list__key summary-key">Court or hearing centre</dt>
+        <dd class="govuk-summary-list__value">{{courtOrHearingCentre | titlecase}}</dd>
+      </div>
+
+      <div class="govuk-summary-list__row" *ngIf="accessType">
+        <dt class="govuk-summary-list__key summary-key">Access</dt>
+        <dd class="govuk-summary-list__value">{{accessType | titlecase}}</dd>
+      </div>
+    </dl>
+  </div>
+
+  <button class="govuk-button govuk-!-margin-right-3">Request access</button>
+</form>
+<a class="govuk-body" (click)="onCancel()" href="javascript:void(0)">Cancel</a>
