+++ src/shared/components/case-viewer/case-basic-access-view/case-basic-access-view.component.ts
@@ -1,0 +1,51 @@
+import { Component, Input, OnDestroy, OnInit } from '@angular/core';
+import { Subscription } from 'rxjs';
+import { CasesService } from '../..';
+import { CaseView } from '../../../domain';
+
+@Component({
+  selector: 'ccd-case-basic-access-view',
+  templateUrl: 'case-basic-access-view.component.html'
+})
+export class CaseBasicAccessViewComponent implements OnInit, OnDestroy {
+
+  @Input()
+  public caseDetails: CaseView = null;
+
+  @Input()
+  public accessType: string = null;
+
+  public courtOrHearingCentre: string = null;
+  private courtOrHearingCentreSubscription: Subscription;
+
+  constructor(
+    private readonly casesService: CasesService
+  ) {}
+
+  public ngOnInit(): void {
+    const locationId = this.caseDetails &&
+                        this.caseDetails.basicFields &&
+                        this.caseDetails.basicFields.caseManagementLocation &&
+                        this.caseDetails.basicFields.caseManagementLocation.baseLocation ?
+      this.caseDetails.basicFields.caseManagementLocation.baseLocation : null;
+
+    if (locationId) {
+      this.courtOrHearingCentreSubscription = this.casesService.getCourtOrHearingCentreName(locationId).subscribe(courtOrHearingCentre =>
+        this.courtOrHearingCentre = courtOrHearingCentre[0] && courtOrHearingCentre[0].building_location_name ?
+                                    courtOrHearingCentre[0].building_location_name :
+                                    null
+      );
+    }
+  }
+
+  public ngOnDestroy(): void {
+    if (this.courtOrHearingCentreSubscription) {
+      this.courtOrHearingCentreSubscription.unsubscribe();
+    }
+  }
+
+  public onCancel(): void {
+    // Navigate to the previous page
+    window.history.go(-1);
+  }
+}
