+++ src/shared/components/case-viewer/case-basic-access-view/index.ts
@@ -1,0 +1,1 @@
+export * from './case-basic-access-view.component';
