+++ src/shared/components/case-viewer/case-full-access-view/index.ts
@@ -1,0 +1,1 @@
+export * from './case-full-access-view.component';
