+++ src/shared/components/case-viewer/case-viewer.module.ts
@@ -1,6 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { ReactiveFormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatTabsModule } from '@angular/material';
 import { RouterModule } from '@angular/router';
 import { AlertModule } from '../../../components/banners/alert';
@@ -24,9 +24,11 @@ import { ErrorMessageComponent } from '../error-message/error-message.component'
 import { ErrorsModule } from '../error/errors.module';
 import { EventTriggerModule } from '../event-trigger';
 import { ComplexModule, EventLogModule, PaletteModule } from '../palette';
+import { CaseBasicAccessViewComponent } from './case-basic-access-view/case-basic-access-view.component';
 import { CaseChallengedAccessRequestComponent } from './case-challenged-access-request/case-challenged-access-request.component';
 import { CaseChallengedAccessSuccessComponent } from './case-challenged-access-success/case-challenged-access-success.component';
 import { CaseEventTriggerComponent } from './case-event-trigger';
+import { CaseFullAccessViewComponent } from './case-full-access-view/case-full-access-view.component';
 import { CaseViewComponent } from './case-view';
 import { CaseViewerComponent } from './case-viewer.component';
 import { CasePrinterComponent, PrintUrlPipe } from './printer';
@@ -50,6 +52,7 @@ import { CaseResolver, EventTriggerResolver } from './services';
     EventLogModule,
     MatTabsModule,
     ComplexModule,
+    FormsModule,
     ReactiveFormsModule,
     AlertModule
   ],
@@ -57,6 +60,8 @@ import { CaseResolver, EventTriggerResolver } from './services';
     CaseEventTriggerComponent,
     CasePrinterComponent,
     CaseViewerComponent,
+    CaseBasicAccessViewComponent,
+    CaseFullAccessViewComponent,
     CaseViewComponent,
     PrintUrlPipe,
     CaseChallengedAccessRequestComponent,
