+++ src/shared/components/helpers/index.ts
@@ -1,0 +1,1 @@
+export * from './init-dialog-helper';
