+++ src/shared/domain/case-view/case-view.model.ts
@@ -31,4 +31,10 @@ export class CaseView {
   events: CaseViewEvent[];
   @Type(() => CaseField)
   metadataFields?: CaseField[];
+  basicFields?: {
+    caseNameHmctsInternal?: string,
+    caseManagementLocation?: {
+      baseLocation?: number
+    }
+  }
 }
